2-search_and_recs/requirements.txt
transformers==4.38.2
sentence-transformers==2.5.1
qdrant-client
pandas
Pillow
fastapi
python-multipart
uvicorn[standard]
torchvision